Multimodal: Rename sample.py to image.py
