+++ lightly_studio/src/lightly_studio/api/db_tables.py
@@ -12,11 +12,11 @@ from lightly_studio.models.dataset import (
 from lightly_studio.models.embedding_model import (
     EmbeddingModelTable,  # noqa: F401, required for SQLModel to work properly
 )
+from lightly_studio.models.image import (
+    ImageTable,  # noqa: F401, required for SQLModel to work properly
+)
 from lightly_studio.models.metadata import (
     SampleMetadataTable,  # noqa: F401, required for SQLModel to work properly
-)
-from lightly_studio.models.sample import (
-    ImageTable,  # noqa: F401, required for SQLModel to work properly
 )
 from lightly_studio.models.sample_embedding import (
     SampleEmbeddingTable,  # noqa: F401, required for SQLModel to work properly
