+++ lightly_studio/src/lightly_studio/api/routes/api/sample.py
@@ -16,7 +16,7 @@ from lightly_studio.api.routes.api.status import (
 from lightly_studio.api.routes.api.validators import Paginated
 from lightly_studio.db_manager import SessionDep
 from lightly_studio.models.dataset import DatasetTable
-from lightly_studio.models.sample import (
+from lightly_studio.models.image import (
     ImageCreate,
     ImageTable,
     ImageView,
