+++ lightly_studio/src/lightly_studio/core/dataset.py
@@ -34,7 +34,7 @@ from lightly_studio.models.annotation.annotation_base import (
     AnnotationType,
 )
 from lightly_studio.models.dataset import DatasetCreate, DatasetTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.resolvers import (
     dataset_resolver,
     embedding_model_resolver,
