+++ lightly_studio/src/lightly_studio/dataset/loader.py
@@ -42,7 +42,7 @@ from lightly_studio.dataset.embedding_manager import (
 from lightly_studio.models.annotation.annotation_base import AnnotationCreate
 from lightly_studio.models.annotation_label import AnnotationLabelCreate
 from lightly_studio.models.dataset import DatasetCreate, DatasetTable
-from lightly_studio.models.sample import ImageCreate, ImageTable
+from lightly_studio.models.image import ImageCreate, ImageTable
 from lightly_studio.resolvers import (
     annotation_label_resolver,
     annotation_resolver,
