+++ lightly_studio/src/lightly_studio/models/image.py
