+++ lightly_studio/src/lightly_studio/resolvers/annotation_resolver/count_annotations_by_dataset.py
@@ -10,7 +10,7 @@ from lightly_studio.models.annotation.annotation_base import (
     AnnotationBaseTable,
 )
 from lightly_studio.models.annotation_label import AnnotationLabelTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.models.tag import TagTable
