+++ lightly_studio/src/lightly_studio/resolvers/annotations/annotations_filter.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel, Field
 from sqlmodel import col
 
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable, AnnotationType
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.models.tag import TagTable
 from lightly_studio.type_definitions import QueryType
