+++ lightly_studio/src/lightly_studio/resolvers/sample_embedding_resolver.py
@@ -7,7 +7,7 @@ from uuid import UUID
 from sqlalchemy import String, cast
 from sqlmodel import Session, col, select
 
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.models.sample_embedding import (
     SampleEmbeddingCreate,
     SampleEmbeddingTable,
