+++ lightly_studio/src/lightly_studio/type_definitions.py
@@ -7,7 +7,7 @@ from uuid import UUID
 from sqlmodel.sql.expression import SelectOfScalar
 
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 
 # Generic query type for filters that work with both data queries and count queries
 QueryType = TypeVar(
