+++ lightly_studio/tests/api/routes/api/annotations/test_create_annotation.py
@@ -8,7 +8,7 @@ from pytest_mock import MockerFixture
 from lightly_studio.api.routes.api.status import HTTP_STATUS_OK
 from lightly_studio.models.annotation.annotation_base import AnnotationType, AnnotationView
 from lightly_studio.models.annotation_label import AnnotationLabelTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.services import annotations_service
 from lightly_studio.services.annotations_service.create_annotation import AnnotationCreateParams
