+++ lightly_studio/tests/core/test_dataset__labelformat.py
@@ -16,7 +16,7 @@ from sqlmodel import select
 from lightly_studio import Dataset
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable
 from lightly_studio.models.annotation_label import AnnotationLabelTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 
 
 class TestDataset:
