+++ lightly_studio/tests/dataset/test_embedding_manager.py
@@ -16,7 +16,7 @@ from lightly_studio.dataset.embedding_manager import (
 )
 from lightly_studio.models.dataset import DatasetTable
 from lightly_studio.models.embedding_model import EmbeddingModelCreate, EmbeddingModelTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 from lightly_studio.models.sample_embedding import SampleEmbeddingTable
 from lightly_studio.resolvers import embedding_model_resolver
 from tests.helpers_resolvers import create_dataset
