+++ lightly_studio/tests/dataset/test_loader__labelformat.py
@@ -16,7 +16,7 @@ from sqlmodel import select
 from lightly_studio.dataset.loader import DatasetLoader
 from lightly_studio.models.annotation.annotation_base import AnnotationBaseTable
 from lightly_studio.models.annotation_label import AnnotationLabelTable
-from lightly_studio.models.sample import ImageTable
+from lightly_studio.models.image import ImageTable
 
 
 def get_input(
